requirements.txt
google-generativeai
psutil
tiktoken
